+++ Example/Controllers/Extensions/ViewController+MiniAppUserInfoProtocol.swift
@@ -15,4 +15,8 @@ extension ViewController {
         }
         return userProfilePhoto
     }
+
+    func getAccessToken(miniAppId: String, completionHandler: @escaping (Result<MATokenInfo, MASDKCustomPermissionError>) -> Void) {
+        completionHandler(.success(.init(accessToken: "ACCESS_TOKEN", expirationDate: Date())))
+    }
 }