+++ MiniApp/Classes/JavascriptBridge/MiniAppScriptMessageHandler.swift
@@ -66,6 +66,8 @@ internal class MiniAppScriptMessageHandler: NSObject, WKScriptMessageHandler {
             fetchProfilePhoto(callbackId: callbackId)
         case .setScreenOrientation:
             setScreenOrientation(requestParam: requestParam, callbackId: callbackId)
+        case .getAccessToken:
+            fetchTokenDetails(callbackId: callbackId)
         }
     }
 
@@ -201,6 +203,21 @@ internal class MiniAppScriptMessageHandler: NSObject, WKScriptMessageHandler {
             executeJavaScriptCallback(responseStatus: .onSuccess, messageId: callbackId, response: MASDKProtocolResponse.success.rawValue)
         } else {
             self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(MiniAppJavaScriptError.valueIsEmpty))
+        }
+    }
+
+    func fetchTokenDetails(callbackId: String) {
+        hostAppMessageDelegate?.getAccessToken(miniAppId: self.miniAppId) { (result) in
+            switch result {
+            case .success(let responseMessage):
+                guard let jsonResponse = ResponseEncoder.encode(data: responseMessage) else {
+                    self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(MiniAppJavaScriptError.internalError))
+                    return
+                }
+                self.executeJavaScriptCallback(responseStatus: .onSuccess, messageId: callbackId, response: jsonResponse)
+            case .failure(let error):
+                self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(error))
+            }
         }
     }
 }