+++ MiniApp/Classes/Protocols/MiniAppUserInfoDelegate.swift
@@ -8,6 +8,9 @@ public protocol MiniAppUserInfoDelegate: class {
 
     /// Interface that is used to retrieve the Image URI
     func getProfilePhoto() -> String?
+
+    /// Interface that is used to retrieve the Token Info
+    func getAccessToken(miniAppId: String, completionHandler: @escaping (Result<MATokenInfo, MASDKCustomPermissionError>) -> Void)
 }
 
 public extension MiniAppUserInfoDelegate {
@@ -17,5 +20,26 @@ public extension MiniAppUserInfoDelegate {
 
     func getProfilePhoto() -> String? {
         return nil
+    }
+
+    func getAccessToken(miniAppId: String, completionHandler: @escaping (Result<MATokenInfo, MASDKCustomPermissionError>) -> Void) {
+        completionHandler(.failure(.failedToConformToProtocol))
+    }
+}
+
+public class MATokenInfo: Codable {
+    let token: String
+    let validUntil: Int
+
+    public init(accessToken: String, expirationDate: Date) {
+        self.token = accessToken
+        self.validUntil = expirationDate.dateToNumber()
+    }
+}
+
+extension Date {
+    func dateToNumber() -> Int {
+        let timeSince1970 = self.timeIntervalSince1970
+        return Int(timeSince1970 * 1000)
     }
 }